closes #151
